Problem: Need dedicated map for sensors
Solution: Create it, and modify functions

Notes:
* this is a preliminary effort for addressing "sub device", i.e.
devices served by data structures of other devices. For example,
sensors data are served by the ambient collection, which is embedded
by the device to which the sensor is connected to.
* The issue with current mapping functions is that it is considered
to be a 1 to 1 mapping, even for the indexed collection. For example,
input.voltage is turned into voltage.input. And outlet.N.voltage is
turned into voltage.outlet.N.
However, for indexed sensors, ambient.N.temperature has to be turned
unindexed, so turned into ambient.temperature.
This is currently not supported, and the present commit may not be
finalized (memory failing!)

Signed-off-by: Arnaud Quette <ArnaudQuette@Eaton.com>